2022/1: split code into modules
Ok, in order to grasp what Rust modules are, I devided the main.rs code in
three different modules, being the main.rs the root, utils.rs the one
containing the read_lines() fn, and elves.rs with all the definitions
(struct and trait) to work with an elf. Considering the complexity of the
code it's really unnecessary, but I'm doing it as a learning exercise.