2022/2: code for the puzzle 2 part 2
Our code had to be changed to lookup for the correct hand based on the other
player's hand and the predefined outcome. We implemented a new game rules
logic so we can lookup for what we want and kept the previous code for
allowing calculations on our hand and outcome for the final puzzle result.