2022/3: code for puzzle 3 part 1
We followed somewhat the same idea we used on puzzle 2, but using different
data types to make sure we learn new stuff. Initially I thought about using
slices for the Compartments members, but the lifetime was not being honorred
in the way we were implementing the overall code, thus I moved to
heap-allocated data types (Vec<> and Box<>).