2022/3: rework code with support for part 1 and 2
This code was re-engineered with other aspects in mind: to some extent,
avoid over-playing with basic data types like arrays or Box<>, and go
directly to Vec<> that the compiler might be able to optimize it better;
on the other hand, play more with more advanced features in order to
get the hand dirty on the borrow checker, like Rc, Weak and RefCells.

The use of these features are totally unecessary to get to the final result
of the puzzle, but they were used as a learning exercise.